+++ src/_pytask/logging.py
@@ -1,3 +1,4 @@
+"""Add general logging capabilities."""
 import platform
 import sys
 from typing import List
@@ -10,6 +11,7 @@ from _pytask.config import hookimpl
 
 @hookimpl
 def pytask_log_session_header(session):
+    """Log the header of a pytask session."""
     tm_width = session.config["terminal_width"]
 
     click.echo(f"{{:=^{tm_width}}}".format(" Start pytask session "), nl=True)
@@ -23,11 +25,14 @@ def pytask_log_session_header(session):
 
     plugin_info = session.config["pm"].list_plugin_distinfo()
     if plugin_info:
-        formatted_plugins_w_versions = ", ".join(_plugin_nameversions(plugin_info))
+        formatted_plugins_w_versions = ", ".join(
+            _format_plugin_names_and_versions(plugin_info)
+        )
         click.echo(f"Plugins: {formatted_plugins_w_versions}")
 
 
-def _plugin_nameversions(plugininfo) -> List[str]:
+def _format_plugin_names_and_versions(plugininfo) -> List[str]:
+    """Format name and version of loaded plugins."""
     values = []  # type: List[str]
     for _, dist in plugininfo:
         # Gets us name and version!