+++ src/pytask/__init__.py
@@ -1,6 +1,7 @@
-from _pytask.cli import main
+from _pytask.build import main
+from _pytask.cli import cli
 from _pytask.config import hookimpl
 from _pytask.mark import MARK_GEN as mark  # noqa: N811
 
-__all__ = ["hookimpl", "main", "mark", "pytask"]
+__all__ = ["cli", "hookimpl", "main", "mark", "pytask"]
 __version__ = "0.0.5"