+++ tests/test_resolve_dependencies.py
@@ -4,12 +4,11 @@ import attr
 import networkx as nx
 import pytest
 from _pytask.exceptions import NodeNotFoundError
-from _pytask.exceptions import ResolvingDependenciesError
 from _pytask.nodes import MetaNode
 from _pytask.nodes import MetaTask
 from _pytask.resolve_dependencies import _check_if_root_nodes_are_available
 from _pytask.resolve_dependencies import pytask_resolve_dependencies_create_dag
-from pytask import main
+from pytask import cli
 
 
 @attr.s
@@ -66,7 +65,8 @@ def test_check_if_root_nodes_are_available():
         _check_if_root_nodes_are_available(dag)
 
 
-def test_cycle_in_dag(tmp_path):
+@pytest.mark.end_to_end
+def test_cycle_in_dag(tmp_path, runner):
     source = """
     import pytask
     from pathlib import Path
@@ -84,9 +84,7 @@ def test_cycle_in_dag(tmp_path):
     """
     tmp_path.joinpath("task_dummy.py").write_text(textwrap.dedent(source))
 
-    session = main({"paths": tmp_path})
+    result = runner.invoke(cli, [tmp_path.as_posix()])
 
-    assert session.exit_code == 3
-    assert isinstance(
-        session.resolving_dependencies_report.exc_info[1], ResolvingDependenciesError,
-    )
+    assert result.exit_code == 4
+    assert "Errors while resolving dependencies" in result.output